Revert "♻️ Refactor chat message (1)"
This reverts commit c9dedd6ac867c7a4407cd9a545cd027f3e4073e9.
