+++ lib/widgets/widgets.dart
@@ -5,4 +5,3 @@ export 'sendMessage.dart';
 export 'poweredBy.dart';
 export 'requireEmailUpfront.dart';
 export 'alert.dart';
-export 'timeWidget.dart';
